+++ wechatkit/__init__.py
@@ -1,0 +1,7 @@
+"""wechatkit init."""
+
+from .api import WechatAPI
+
+__all__ = [
+    'WechatAPI'
+]