+++ wechatkit/user.py
@@ -1,7 +1,7 @@
 """wechatkit user"""
-from wechatkit import consts
-from wechatkit.utils import RequestUtil
-from wechatkit.exceptions import WechatException
+from . import consts
+from .utils import RequestUtil
+from .exceptions import WechatException
 
 
 class WechatUserAPI(object):